+++ Euro_Denominator_Backend/src/main/java/org/example/euro_denominator/model/DenominationResponse.java
@@ -7,22 +7,22 @@ import java.util.Map;
  * Contains the computed denomination breakdown and its difference from the previous result.
  */
 public class DenominationResponse {
-    private Map<Integer, Integer> breakdown;
-    private Map<Integer, Integer> differenceFromPrevious;
+    private Map<Integer, Long> breakdown;
+    private Map<Integer, Long> differenceFromPrevious;
 
-    public Map<Integer, Integer> getBreakdown() {
+    public Map<Integer, Long> getBreakdown() {
         return breakdown;
     }
 
-    public void setBreakdown(Map<Integer, Integer> breakdown) {
+    public void setBreakdown(Map<Integer, Long> breakdown) {
         this.breakdown = breakdown;
     }
 
-    public Map<Integer, Integer> getDifferenceFromPrevious() {
+    public Map<Integer, Long> getDifferenceFromPrevious() {
         return differenceFromPrevious;
     }
 
-    public void setDifferenceFromPrevious(Map<Integer, Integer> differenceFromPrevious) {
+    public void setDifferenceFromPrevious(Map<Integer, Long> differenceFromPrevious) {
         this.differenceFromPrevious = differenceFromPrevious;
     }
 }
